lab2_redis/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
redis
